+++ Assets/Mateo/Scripts/Editor/ScriptEditor.cs.meta
