+++ Assets/Mateo/Scripts/ScriptableObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea94856313adb234b823c72e7d6ff263
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
