+++ Assets/Mateo/Scripts/ScriptableObject/Script_01.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c4e355d4fef8df2488fd01dccb52f278, type: 3}
+  m_Name: Script_01
+  m_EditorClassIdentifier: 
+  ArrayLeght: 0
+  commands: []
