+++ Assets/Mateo/Scripts/ScriptableObject/Script_01.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b4fa5cd88073704aa8762fa66ec7231
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
